+++ payments-core/src/test/java/com/stripe/android/view/PaymentAuthWebChromeClientTest.kt
@@ -2,7 +2,7 @@ package com.stripe.android.view
 
 import android.webkit.ConsoleMessage
 import com.google.common.truth.Truth.assertThat
-import com.stripe.android.FakeLogger
+import com.stripe.android.testing.FakeLogger
 import org.junit.runner.RunWith
 import org.mockito.kotlin.mock
 import org.robolectric.RobolectricTestRunner