+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/paymentdatacollection/ach/USBankAccountFormViewModel.kt
@@ -470,17 +470,9 @@ internal class USBankAccountFormViewModel @Inject internal constructor(
 
     private fun collectBankAccountForIntent(clientSecret: String) {
         val configuration = if (args.instantDebits) {
-            CollectBankAccountConfiguration.InstantDebits(
-                email = email.value,
-                elementsSessionContext = ElementsSessionContext(
-                    linkMode = args.linkMode,
-                ),
-            )
+            createInstantDebitsConfiguration()
         } else {
-            CollectBankAccountConfiguration.USBankAccount(
-                name = name.value,
-                email = email.value,
-            )
+            createUSBankAccountConfiguration()
         }
 
         if (args.isPaymentFlow) {
@@ -498,6 +490,33 @@ internal class USBankAccountFormViewModel @Inject internal constructor(
                 configuration = configuration,
             )
         }
+    }
+
+    private fun createInstantDebitsConfiguration(): CollectBankAccountConfiguration.InstantDebits {
+        val initializationMode = if (args.clientSecret == null) {
+            ElementsSessionContext.InitializationMode.DeferredIntent
+        } else if (args.isPaymentFlow) {
+            ElementsSessionContext.InitializationMode.PaymentIntent(args.stripeIntentId!!)
+        } else {
+            ElementsSessionContext.InitializationMode.SetupIntent(args.stripeIntentId!!)
+        }
+
+        return CollectBankAccountConfiguration.InstantDebits(
+            email = email.value,
+            elementsSessionContext = ElementsSessionContext(
+                initializationMode = initializationMode,
+                amount = args.formArgs.amount?.value,
+                currency = args.formArgs.amount?.currencyCode,
+                linkMode = args.linkMode,
+            ),
+        )
+    }
+
+    private fun createUSBankAccountConfiguration(): CollectBankAccountConfiguration.USBankAccount {
+        return CollectBankAccountConfiguration.USBankAccount(
+            name = name.value,
+            email = email.value,
+        )
     }
 
     private fun collectBankAccountForDeferredIntent() {